New CRD: USBDevicesConfig
This custom resource allows the administrator to identify USB devices
plugged in the cluster's node and allocate them to VMs.

At this moment, the only implemented selector uses vendor:product
information, see the example below.

 > e.g: usb-devices-config.yaml
 >  apiVersion: usb.kubevirt.io/v1alpha1
 >  kind: USBDevicesConfig
 >  metadata:
 >    name: storage
 >    namespace: default
 >  spec:
 >    usb:
 >      - resourceName: kubevirt.io/storage
 >        usbHostDevices:
 >          - selectByVendorProduct: 46f4:0001

The exposition of USB devices and allocation of USBDevicesConfig
resource is described in follow up patches.

Signed-off-by: L. Pivarc <lpivarc@redhat.com>
Signed-off-by: Victor Toso <victortoso@redhat.com>