hooks: Add v1alpha3 with Shutdown method
There is a but in sidecar lifecycle that might keep the sidecar
running even when its Pod has terminated, leaving the Pod dangling
with "NotReady" state. This is being tracked in:

    https://github.com/kubernetes/kubernetes/pull/116429

This bug shows itself in KubeVirt on migration. After the first
migration job, the VM can't be migrated again as previous migration
does not finish.

In the hooks' v1alpha3 we introduce the Shutdown() method to allow
notify the sidecar that it should exit(). Follow up patches will
implement this method in the cmd/example-hook-sidecar.

Fixes: https://github.com/kubevirt/kubevirt/issues/8395
Signed-off-by: Victor Toso <victortoso@redhat.com>