api: core: Add USB to Host Devices
Under KubeVirt CRD we already have a section for Host Devices in which
we currently allow PCI and Mediated type of devices. This patch
adds USB type of devices.

The cluster admin can add selectors to the KubeVirt CRD and that would
allow exporting USB devices plugged into Nodes to be allocated by
users. In the example configuration below, any USB Device with
vendor:product being 46f4:0001 will be exposed under
"kubevirt.io/storage" resource name:

KubeVirt CRD
------------
  spec:
    configuration:
      permittedHostDevices:
        usb:
          - resourceName: kubevirt.io/storage
            selectors:
              - vendor: "46f4"
                product: "0001"

The user that wants to create a VM with said USB device can do so with
the following configuration:

In a VMI
--------
  spec:
    domain:
      devices:
        hostDevices:
        - deviceName: kubevirt.io/storage
          name: usb-storage1

Signed-off-by: Victor Toso <victortoso@redhat.com>