virtwrap: efi: use libvirt firmware auto selection
Since libvirt 8.6.0 we can use firmware auto selection to enable
secure boot. See:

  https://libvirt.org/kbase/secureboot.html

The main change needed is properly set to enabled the firmware's
features 'secure-boot' and 'enrolled-keys' in order to boot with
secure-boot enabled.

The main difference is that by setting bot 'secure-boot' and
'enrolled-keys' is that it denies unsigned Guests to boot. Prior to
this commit, the Guest would boot with Secure Boot disabled.

Signed-off-by: Victor Toso <victortoso@redhat.com>