virtctl: usbredir: error when connection is closed
The expected end of USB redirection is only when the user to stop
sharing the USB devices. If the connection is closed without expecting
it, we should treat it as error.

The follow up patch "virtctl: usbredir: Add Stop() to Client" will
improve this logic to make it clear when we expected it to stop.

Signed-off-by: Victor Toso <victortoso@redhat.com>