tests: Fix enabling ssh password authentication
The current sed script works under the assumption that the
sshd_config file contains the line

  PasswordAuthentication no

That is the case for the current Alpine 3.16 image, though I
haven't been able to figure out where it comes from since it
doesn't appear to exits even in the old Alpine package. The
line is not there in the Alpine 3.19 image, which means that
the sed script does nothing and password authentication
remains disabled, leading to a bunch of failed test cases.

Anyway, cloud-init implements a native way to enable password
authentication, so we can just switch to that one and make
things both simpler and more robust.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>