hooks: move the per-version callback to its own method
This allow us to have similar logic (e.g: equal to OnDefineDomain)
between all gRPC methods which in turn helps to find issues with the
logic.

I've now noticed that the PreCloudInitIso() is only called for the
very first callback in the list: A non tested bug in case we have
multiple sidecars with that function.

I've also noticed that for the Shutdown(), in case of errors, it'll
not call further callbacks. This too is a non tested bug in case we
have multiple sidecars. Note that not calling Shutdown() is an issue
for Migration.

As this and follow-up commit goals are only move code around without
logic change, I've added two FIXMEs for each issue to make sure this
is not ignored.

Signed-off-by: Victor Toso <victortoso@redhat.com>